Rename resources_api.rb -> resource_api.rb
This way the file name matches the module name
